+++ for_time_measure/algorithms.h
@@ -6,12 +6,12 @@
 
 /// Dijkstra's algorithm for finding the shortest paths from a single node to all other nodes in a weighted graph.
 /// \param graph - adjacency matrix of the graph
-void dijkstra(const std::vector<std::vector<int>>& graph);
+void dijkstra(const std::vector<std::vector<int64_t>>& graph);
 
 /// Floyd-Warshall algorithm for finding the shortest paths between all pairs of vertices in a weighted graph.
 /// \param graph - adjacency matrix of the graph
-void floydWarshall(const std::vector<std::vector<int>>& graph);
+void floydWarshall(const std::vector<std::vector<int64_t>>& graph);
 
 /// Bellman-Ford algorithm for finding the shortest paths from a single node to all other nodes in a weighted graph.
 /// \param graph - adjacency matrix of the graph
-void bellmanFord(const std::vector<std::vector<int>>& graph);
+void bellmanFord(const std::vector<std::vector<int64_t>>& graph);
