+++ for_time_measure/dijkstra.cpp
@@ -3,19 +3,19 @@
 #include <queue>
 #include <climits>
 
-void dijkstra(const std::vector<std::vector<int>>& graph) {
-    int n = graph.size();
-    int source = 0;
-    std::vector<int> dist(n, INT_MAX);
+void dijkstra(const std::vector<std::vector<int64_t>>& graph) {
+    int64_t n = graph.size();
+    int64_t source = 0;
+    std::vector<int64_t> dist(n, INT64_MAX);
     std::vector<bool> visited(n, false);
 
     dist[source] = 0;
 
-    std::priority_queue<std::pair<int, int>, std::vector<std::pair<int, int>>, std::greater<std::pair<int, int>>> pq;
+    std::priority_queue<std::pair<int64_t, int64_t>, std::vector<std::pair<int64_t, int64_t>>, std::greater<std::pair<int64_t, int64_t>>> pq;
     pq.push(std::make_pair(0, source));
 
     while (!pq.empty()) {
-        int u = pq.top().second;
+        int64_t u = pq.top().second;
         pq.pop();
 
         if (visited[u]) {
@@ -24,10 +24,10 @@ void dijkstra(const std::vector<std::vector<int>>& graph) {
 
         visited[u] = true;
 
-        for (int v = 0; v < n; ++v) {
-            int weight = graph[u][v];
+        for (int64_t v = 0; v < n; ++v) {
+            int64_t weight = graph[u][v];
 
-            if (!visited[v] && dist[u] != INT_MAX && dist[u] + weight < dist[v]) {
+            if (!visited[v] && dist[u] != INT64_MAX && dist[u] + weight < dist[v]) {
                 dist[v] = dist[u] + weight;
                 pq.push(std::make_pair(dist[v], v));
             }
